Change transfer size type from uint32_t to size_t for >4GB support (#2444)
### Issue

Transfers larger than 4GB overflow uint32_t size parameters.

### Description

Change transfer size type from `uint32_t` to `size_t` for
`write_to_device`, `read_from_device`, and TlbWindow
`read/write_block_reconfigure` across the full class hierarchy:
- `Chip` interface and all implementations (Local, Remote, Mock,
Simulation)
- `TTDevice` and all subclasses
- `DeviceProtocol` and all implementations (PCIe, Remote, JTAG)
- `TlbWindow` reconfigure methods and internal loop variables
- `Cluster` public API
- Python nanobind bindings (removed narrowing `static_cast<uint32_t>`
casts)

DMA functions already used `size_t` — no changes needed there.

### Testing

CI

### API Changes

Public API size parameters changed from `uint32_t` to `size_t` for
`write_to_device`, `read_from_device`, and TlbWindow reconfigure
methods. This is source-compatible for callers passing `uint32_t` values
(implicit widening conversion).

+++ device/api/umd/device/cluster.hpp
@@ -360,7 +360,7 @@ public:
      * @param core Core to target.
      * @param addr Address to write to.
      */
-    void write_to_device(const void* mem_ptr, uint32_t size_in_bytes, ChipId chip, CoreCoord core, uint64_t addr);
+    void write_to_device(const void* mem_ptr, size_t size_in_bytes, ChipId chip, CoreCoord core, uint64_t addr);
 
     /**
      * Read uint32_t data from a specified device, core and address to host memory (defined for Silicon).
@@ -373,7 +373,7 @@ public:
      * @param addr Address to read from.
      * @param size Number of bytes to read.
      */
-    void read_from_device(void* mem_ptr, ChipId chip, CoreCoord core, uint64_t addr, uint32_t size);
+    void read_from_device(void* mem_ptr, ChipId chip, CoreCoord core, uint64_t addr, size_t size);
 
     /**
      * Write uint32_t data (as specified by ptr + len pair) to specified device, core and address (defined for Silicon).
